Update wp-env.ts to be more flexible
