+++ config/wp-env/run.ts
@@ -1,0 +1,3 @@
+import { WpEnv } from './WpEnv';
+
+new WpEnv();
