+++ package.json
@@ -36,7 +36,8 @@
 		"test:e2e": "jest --config packages/e2e-tests/jest.playwright.config.ts",
 		"test:e2e:clearCache": "jest --config packages/e2e-tests/jest.playwright.config.ts --clearCache",
 		"test": "node scripts/test.js",
-		"docker:make-config": "yarn ts-node config/wp-env.ts",
+		"docker:cli": "yarn ts-node ./config/wp-env/run.ts run",
+		"docker:make-config": "yarn ts-node ./config/wp-env/run.ts make config",
 		"docker:start": "yarn docker:make-config && yarn wp-env start",
 		"docker:update": "yarn docker:start --update",
 		"docker:stop": "yarn wp-env stop",
